Alteração do campo timesPerWeek para Double e adicionando validaçoes no Exercise

+++ src/main/java/com/example/labmedical/controller/dtos/response/ExerciseResponse.java
@@ -15,7 +15,7 @@ public class ExerciseResponse {
     private LocalTime timeCreated;
     @Enumerated(EnumType.STRING)
     private ExerciseType exerciseType;
-    private Integer timesPerWeek;
+    private Double timesPerWeek;
     private String exerciseSeriesName;
     private Long patientId;
     private boolean status = true;
